debug

+++ src/utils/cloudinary.js
@@ -1,27 +1,34 @@
-import { v2 as cloudinary } from 'cloudinary'
-import fs from 'fs'
-
-
-
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_NAME, 
+import { v2 as cloudinary } from "cloudinary"
+import fs from "fs"
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET
-});
+})
 
-const uplodCloudinary = async(localFilePath)=>{
-    try {
-        if(!localFilePath) return null;
-     const response =  await cloudinary.uploader.upload(localFilePath,{
-            resource_type:"auto"
-            
-        })
-        console.log("file uploaded sucessfully ", response.url)
-        return response
-    } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
+
+const uplodCloudinary = async (localFilePath) => {
+  try {
+    if (!localFilePath) return null
+
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto"
+    })
+     fs.unlinkSync(localFilePath)
+    //console.log("File uploaded successfully:", response.url)
+
+    return response
+
+  } catch (error) {
+
+    console.log("Cloudinary error:", error)
+
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath)
     }
+
+    return null
+  }
 }
 
-export {uplodCloudinary}
+export { uplodCloudinary }
